Cambio tipo de dato a parametros del método sumar

+++ demo-git/src/Math.java
@@ -1,7 +1,7 @@
 
 public class Math {
-	public static int sumar(int numero1,int numero2){
-		return numero1+numero2;
+	public static int sumar(short numero1, short numero2){
+		return numero1 + numero2;
 	}
 	
 	public static int multiplicar(int numero1,int numero2){
